[serialization] minor
* moved load, list_models functions to ext src.
* Added nicer sphinx doc (wip)
