+++ doc/source/api/index_serialization.rst
@@ -1,0 +1,2 @@
+
+.. automodule:: pyemma._base.serialization
