+++ pyemma/__init__.py
@@ -36,44 +36,7 @@ from . import util
 from . import plots
 from . import thermo
 
-
-def load(filename, model_name='default'):
-    """ Restores a previously saved model or estimator from disk.
-
-    Parameters
-    ----------
-    filename : str
-        path to filename, where the model has been stored.
-    model_name: str, default='default'
-        if multiple models are contained in the file, these can be accessed by
-        their name. Use func:`pyemma.list_models` to get a representation of all stored models.
-
-    Returns
-    -------
-    obj : Model or Estimator
-        the instance containing the same parameters as the saved model/estimator.
-
-    """
-    from ._base.serialization.serialization import SerializableMixIn
-    return SerializableMixIn.load(file_name=filename, model_name=model_name)
-
-
-def list_models(filename):
-    """ Lists all models in given filename.
-    Parameters
-    ----------
-    filename: str
-        path to filename, where the model has been stored.
-
-    Returns
-    -------
-    obj: dict
-        A mapping by name and a comprehensive description like this:
-        {model_name: {'repr' : 'string representation, 'created': 'human readable date', ...}
-    """
-    from ._base.serialization.h5file import H5File
-    with H5File(filename, mode='r') as f:
-        return f.models_descriptive
+from ._base.serialization import load, list_models
 
 
 def _version_check(current, testing=False):
